valoresespeciais

+++ semana2/valoresespeciais/Program.cs
@@ -11,18 +11,18 @@ namespace valoresespeciais
 
             Console.WriteLine();
 
-            Console.WriteLine(short.MaxValue);
-            Console.WriteLine(short.MinValue);
+            Console.WriteLine(sbyte.MaxValue);
+            Console.WriteLine(sbyte.MinValue);
 
             Console.WriteLine();
 
-            Console.WriteLine(short.MaxValue);
-            Console.WriteLine(short.MinValue);
+            Console.WriteLine(long.MaxValue);
+            Console.WriteLine(long.MinValue);
 
             Console.WriteLine();
 
-            Console.WriteLine(short.MaxValue);
-            Console.WriteLine(short.MinValue);
+            Console.WriteLine(int.MaxValue);
+            Console.WriteLine(int.MinValue);
 
             Console.WriteLine();
